New logging code.  The idea is that instead of manually checking whether a log line is necessary, then calling an appropriate log function, you now call one unified irclog_log() function.
The arguments are the event you wish to log (one of the IRC_LOG_* macros),
a destination log file (channel name, nickname, IRC_LOGFILE_SERVER or
IRC_LOGFILE_ALL), a source (nick!user@host) and the text.

Huge differences to the on-disk log files as well, the internal logs in /tmp
now have the format:
	timestamp event dest source text
eg.
	1034880210 message #foo Keybuk!scott@nethack.netsplit.com Hello there
or a private conversation:
	1034880233 message Paul Keybuk!scott@nethack.netsplit.com Hey Paul!
	1034880253 message Paul Paul!paulc@somewhere.com rah!
(Note how messages-from-us are now stored - much neater :)

The log files for the user are in a "sensible output" format, kinda like
a screendump of an IRC client.  If log_timestamp is yes then they have
the timestamp on the front in the form [1 Mon 12:00]